feat: move rename function into class
